Call long overload in MakeItem to fix recursion
In Kahuna.Tests/Server/TestPersistenceBackends.cs the MakeItem(string,int) helper was accidentally calling itself, causing infinite recursion. Update the call to delegate to the overload that accepts a long (MakeItem(key, (long)i)), preventing a stack overflow when constructing test items.

+++ Kahuna.Tests/Server/TestPersistenceBackends.cs
@@ -22,7 +22,7 @@ public class TestPersistenceBackends
             lastModifiedCounter: 0,
             state: (int)KeyValueState.Set);
 
-    private static PersistenceRequestItem MakeItem(string key, int i) => MakeItem(key, i);
+    private static PersistenceRequestItem MakeItem(string key, int i) => MakeItem(key, (long)i);
 
     private static List<PersistenceRequestItem> MakeItems(string prefix, int count, int startIndex = 0) =>
         Enumerable.Range(startIndex, count)
